requirements.txt
# LLM (resumos + provas)
openai==1.102.0
python-dotenv==1.1.0

# Scraping das transcrições (Alura)
playwright==1.51.0
beautifulsoup4==4.13.4
tqdm==4.66.5
requests==2.32.3
